// src/app/Core/Interfaces/iitems-detailes.ts
export interface IitemsDetailes {
  item: Item;
}

export interface Item {
  Code: string;
  GroupGuid: string;
  ItemGroupGuid: string;
  ItemGroupCode: string;
  ItemGroupId: number;
  ItemType: number;
  RecipeItemType: number;
  RecipeItems: any[];
  CouponItems: null;
  NotTaxable: boolean;
  NotDiscountable: boolean;
  ExpiryDateRequired: boolean;
  MaxLimit: number;
  MinLimit: number;
  Category: string;
  CardImage: string;
  TBL007ID: number;
  SmallestUnit: number;
  TotalInQty: number;
  TotalOutQty: number;
  Balance: number;
  Vat: number;
  ItemUnits: ItemUnit[];
  ChoseUnite: null;
  LessQuantity: number;
  OrderLimit: number;
  greatestQuantity: number;
  DiscountTax: number;
  SerialNo: number;
  TypeOfProfit: number;
  ProfitRate: number;
  valueOfProfit: number;
  ProductCompany: null;
  ProductDescription: null;
  DescriptionNotes: null;
  ModelType: null;
  Model: null;
  IsModelRepeated: boolean;
  IsModelTypeRepeated: boolean;
  SupplierId: number;
  SupplierName: null;
  HomeCountry: null;
  CustomsCode: null;
  CartonSize: null;
  ColorId: number;
  Color: null;
  DealingExpiryDate: boolean;
  HasBonus: boolean;
  IsTobago: boolean;
  TobagoVat: number;
  TobagoVatAmount: number;
  IsPurchase: null;
  IsWorking: boolean;
  Imeges: null;
  ItemPlace: number;
  Description: null;
  HowToUse: null;
  Specifications: null;
  Opinions: null;
  Id: number;
  No: number;
  Guid: string;
  NameAr: string;
  NameEn: string;
  BranchId: number;
  BranchName: null;
  CompanyId: number;
  CompanyName: null;
  DocRefGuid: string;
  IsActive: boolean;
  CreatedDate: Date;
  CreatedBy: null;
  IsUpdated: boolean;
  LastUpdatedBy: null;
  LastUpdatedDate: Date;
  UpdatedCount: number;
  IsCanceled: boolean;
  CancelById: number;
  CancelBy: null;
  Notes: null;
  CancelDate: Date;
  ModifyById: null;
  ModifyByName: string;
  LastModifiedDate: Date;
  ModifyCount: number;
  updateFlag: null;
}

export interface ItemUnit {
  Code: string;
  UnitId: number;
  UnitName: string;
  UnitCode: string;
  ManualCode: string;
  Barcode: string;
  Gtin: string;
  Weight: number;
  Value: number;
  Factor: number;
  TaxRatio: number;
  MaxLimit: number;
  MinLimit: number;
  Price: number;
  DefaultCost: number;
  Vat: number;
  PriceTax: number;
  StanderCost: number;
  StanderCostTax: number;
  SupplierId: number;
  ColorId: number;
  Color: null;
  ItemId: number;
  ItemCode: string;
  ItemGroupId: number;
  ItemGroupName: null;
  ItemGroupGuid: string;
  Qty: number;
  WA: number;
  IsPurchase: null;
  IsWorking: boolean;
  ItemPlace: number;
  Description: null;
  HowToUse: null;
  Specifications: null;
  Opinions: null;
  ModelType: null;
  Model: null;
  ItemImages: ItemImage[];
  ItemInStorageDetails: any[];
  ItemsSpare: any[];
  DealingExpiryDate: boolean;
  IsTobago: boolean;
  TobagoVat: number;
  TobagoVatAmount: number;
  Id: number;
  No: number;
  Guid: string;
  NameAr: string;
  NameEn: null;
  BranchId: number;
  BranchName: null;
  CompanyId: number;
  CompanyName: null;
  DocRefGuid: string;
  IsActive: boolean;
  CreatedDate: Date;
  CreatedBy: null;
  IsUpdated: boolean;
  LastUpdatedBy: null;
  LastUpdatedDate: Date;
  UpdatedCount: number;
  IsCanceled: boolean;
  CancelById: number;
  CancelBy: null;
  Notes: null;
  CancelDate: Date;
  ModifyById: null;
  ModifyByName: null;
  LastModifiedDate: Date;
  ModifyCount: number;
  updateFlag: null;
}